pyproject.toml
[tool.poetry]
name = "examplecompanywebsite"
version = "0.1.0"
description = ""
authors = ["Tomasz Sas <tomaszsas.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.30.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
